Fixed: Beast, Eye Ra, Panorama & Blood Demon
Removed some of the leftover bonus attributes associated with a baseline script that wasn't intended to make it on to the published version of the race.
Also fixe an issue which resulted in Blood Demon never receiving the speed from when the skill were supposed to grant the player speed as an upgrade.